nexp_contracts/pyproject.toml
[tool.poetry]
name = "Nexp_pwc"
version = "0.1.0"
description = "Algorand smart contracts"
authors = [" <>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
beaker-pyteal = "^1.1.1"
algokit-utils = "^2.0.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pip-audit = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

